refactor: remove legacy API service and related code
This commit removes obsolete code related to the legacy API service, including:

- api.go, api_test.go, api_adapter_test.go, api_provider_test.go, api_test_helper.go
- file_adapter_test.go and provider_detection_test.go
- All .go.disabled files and the unused runutil package
- Legacy token management files

Also:
- Kept and updated adapters.go to maintain compatibility
- Added minimal token stubs to keep tests passing

This is part of the R2 codebase optimization plan to eliminate dead code,
reducing complexity while preserving core functionality.
