+++ BACKLOG.md
@@ -1,9 +1,13 @@
 # BACKLOG
 
-- [~] **T13: Add synthesis step**
-  Add a built-in synthesis step where outputs from multiple preceding steps (e.g., multiple model responses, critiques) are sent to a final model for summarization / consolidation. i suspect implementation is some optional --synthesis-model flag where the user can set the model to use for synthesis
+## Top Priorities
 
-- build mcp server
+- [ ] **T25: Develop workflow engine**
+  Develop a flexible multi-step workflow engine allowing users to define arbitrary sequences of operations (e.g., plan -> critique -> revise -> synthesize) via configuration or cli flags.
+  * this might just be achieved by designing the program to support piping, ie take the multiple responses generated by the first pass and pipe them in as context for another pass
+
+- [ ] **T37: Integrate with Claude Code**
+  Investigate presenting `thinktank` itself as a tool to claude code, likely via mcp, defining its capabilities for planning, critique, and refinement
 
 - [ ] **T14: Implement semantic versioning**
   Implement semantic versioning (ideally automatically managed somehow ... conventional commits?)
@@ -60,13 +64,6 @@
 - [ ] **T24: Estimate request costs**
   Estimate request cost based on token count and model pricing information, log the estimated cost, and potentially integrate with provider billing apis for accuracy.
 
-- [ ] **T25: Develop workflow engine**
-  Develop a flexible multi-step workflow engine allowing users to define arbitrary sequences of operations (e.g., plan -> critique -> revise -> synthesize) via configuration or cli flags.
-  * this might just be achieved by designing the program to support piping, ie take the multiple responses generated by the first pass and pipe them in as context for another pass
-
-- [ ] **T26: Support multiple concurrent models**
-  Support querying multiple models concurrently for the same task (`--model model_a,model_b`) to get a "council of experts" perspective, returning labeled outputs.
-
 - [ ] **T27: Implement built-in workflows**
   Implement a built-in plan -> critique -> refine workflow activated by a specific flag (e.g., `--critique-refine`).
 
@@ -96,9 +93,6 @@
 
 - [ ] **T36: Support Claude memory files**
   Make `thinktank` aware of claude code's memory files (`CLAUDE.md`, `CLAUDE.local.md`) to read configuration settings, respecting the same hierarchy
-
-- [ ] **T37: Integrate with Claude Code**
-  Investigate presenting `thinktank` itself as a tool to claude code, potentially via mcp, defining its capabilities for planning, critique, and refinement
 
 - [ ] **T38: Improve error messages**
   Review and significantly improve the clarity, detail, and actionability of all error messages throughout the application
