--- internal/openai/disabled/openai_model_info_test.go.disabled
@@ -1,32 +1,0 @@
-// Package openai provides an implementation of the LLM client for the OpenAI API
-package openai
-
-import (
-	"context"
-	"errors"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-// TestGetModelLimits has been removed as part of token handling removal
-
-// TestGetModelName verifies that the client's GetModelName method
-// correctly returns the model name
-func TestGetModelName(t *testing.T) {
-	// Create a client with a specific model name
-	client := &openaiClient{
-		modelName: "gpt-4",
-	}
-
-	// Get model name
-	modelName := client.GetModelName()
-
-	// Verify result
-	assert.Equal(t, "gpt-4", modelName, "GetModelName should return the correct model name")
-}
-
-// TestGetModelLimitsWithUnknownModel has been removed as part of token handling removal
-
-// TestGetModelLimitsIntegrationWithMockProvider has been removed as part of token handling removal
