--- internal/runutil/README.md
@@ -1,7 +1,0 @@
-# Run Utilities
-
-This directory previously contained utilities for generating random adjective-noun combinations for run names.
-
-As of the directory naming simplification project, this functionality has been replaced with a timestamp-based naming system. The functionality has been moved to the `internal/thinktank/app.go` file.
-
-This directory is kept for compatibility reasons but will be removed in a future cleanup.
